Refactor ClientUnaryCallImpl
